--- app/src/main/java/com/nimontoy/android/helper/socialLogin/SocialLoginHelper.kt
@@ -1,67 +1,0 @@
-package com.nimontoy.android.helper.socialLogin
-
-import android.util.Log
-import android.widget.Toast
-import com.facebook.AccessToken
-import com.facebook.CallbackManager
-import com.facebook.FacebookCallback
-import com.facebook.FacebookException
-import com.facebook.login.LoginManager
-import com.facebook.login.LoginResult
-import com.google.firebase.auth.FacebookAuthProvider
-import com.google.firebase.auth.FirebaseAuth
-
-class SocialLoginHelper {
-    private lateinit var callbackManager: CallbackManager
-    private val TAG = "LoginActivity"
-    private lateinit var auth : FirebaseAuth
-
-    fun FacebookLoginHelper() {
-
-        callbackManager = CallbackManager.Factory.create();
-
-        LoginManager.getInstance().registerCallback(callbackManager,
-            object : FacebookCallback<LoginResult> {
-                override fun onSuccess(loginResult: LoginResult) {
-                    // App code
-                    Log.d(TAG, "onSucces LoginResult=$loginResult")
-                    //handleFacebookAccessToken(loginResult.accessToken)
-                }
-
-                override fun onCancel() {
-                    // App code
-                    Log.d(TAG, "onCancel")
-                }
-
-                override fun onError(exception: FacebookException) {
-                    // App code
-                    Log.d(TAG, "onError")
-                }
-            })
-
-        /*
-        fun handleFacebookAccessToken(token: AccessToken) {
-            Log.d(TAG, "handleFacebookAccessToken:$token")
-
-            val credential = FacebookAuthProvider.getCredential(token.token)
-            auth.signInWithCredential(credential)
-                .addOnCompleteListener(this) { task ->
-                    if (task.isSuccessful) {
-                        // Sign in success, update UI with the signed-in user's information
-                        Log.d(TAG, "signInWithCredential:success")
-                        val user = auth.currentUser
-                        updateUI(user)
-                    } else {
-                        // If sign in fails, display a message to the user.
-                        Log.w(TAG, "signInWithCredential:failure", task.exception)
-                        Toast.makeText(baseContext, "Authentication failed.",
-                            Toast.LENGTH_SHORT).show()
-                        updateUI(null)
-                    }
-
-                }
-        }
-        */
-
-    }
-}
